busstop update

+++ de.bund.bfr.busstop/src/de/bund/bfr/busstopp/dao/Dao.java
@@ -33,10 +33,13 @@ public enum Dao {
 				for (File path : paths) {				
 					//System.out.println(path);
 					if (path.isDirectory()) {
-						long l = Long.parseLong(path.getName());
-						ItemLoader item = new ItemLoader(l, path);
-						if (!item.isDeleted()) contentProvider.put(l, item);	
-						else contentDelProvider.put(l, item);
+						String pn = path.getName();
+						if (!pn.startsWith("out_")) {
+							long l = Long.parseLong(pn);
+							ItemLoader item = new ItemLoader(l, path);
+							if (!item.isDeleted()) contentProvider.put(l, item);	
+							else contentDelProvider.put(l, item);
+						}
 					}
 				}
 			}
@@ -65,11 +68,14 @@ public enum Dao {
 				for (File path : paths) {				
 					//System.out.println(path);
 					if (path.isDirectory()) {
-						long l = Long.parseLong(path.getName());
-						ItemLoader item = new ItemLoader(l, path);
-						if (item.isDeleted()) { //  || item.getXml().getIn().getFilename().indexOf(":") >= 0
-							contentDelProvider.remove(l);
-							if (deleteDir(path)) result++;
+						String pn = path.getName();
+						if (!pn.startsWith("out_")) {
+							long l = Long.parseLong(pn);
+							ItemLoader item = new ItemLoader(l, path);
+							if (item.isDeleted()) { //  || item.getXml().getIn().getFilename().indexOf(":") >= 0
+								contentDelProvider.remove(l);
+								if (deleteDir(path)) result++;
+							}
 						}
 					}
 				}
@@ -92,13 +98,16 @@ public enum Dao {
 				for (File path : paths) {				
 //					System.out.println(path);
 					if (path.isDirectory()) {
-						long l = Long.parseLong(path.getName());
-						ItemLoader item = new ItemLoader(l, path);
-						if (!item.isDeleted()) {
-							item.delete();
-							contentProvider.remove(l);
-							contentDelProvider.put(l, item);
-							result++;
+						String pn = path.getName();
+						if (!pn.startsWith("out_")) {
+							long l = Long.parseLong(pn);
+							ItemLoader item = new ItemLoader(l, path);
+							if (!item.isDeleted()) {
+								item.delete();
+								contentProvider.remove(l);
+								contentDelProvider.put(l, item);
+								result++;
+							}
 						}
 					}
 				}
